:recycle: refactor move monitoring tools to `src`
