+++ api/lib/domain/usecases/create-organizations-with-tags-and-target-profiles.js
@@ -10,10 +10,10 @@ import {
 } from '../../../src/shared/domain/errors.js';
 import { Organization, OrganizationForAdmin, OrganizationTag } from '../../../src/shared/domain/models/index.js';
 import * as codeGenerator from '../../../src/shared/domain/services/code-generator.js';
+import { monitoringTools } from '../../../src/shared/infrastructure/monitoring-tools.js';
 import { PromiseUtils } from '../../../src/shared/infrastructure/utils/promise-utils.js';
 import { CONCURRENCY_HEAVY_OPERATIONS } from '../../infrastructure/constants.js';
 import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
-import { monitoringTools } from '../../infrastructure/monitoring-tools.js';
 
 const SEPARATOR = '_';
