+++ api/lib/domain/usecases/send-shared-participation-results-to-pole-emploi.js
@@ -1,8 +1,8 @@
 import { PoleEmploiSending } from '../../../src/shared/domain/models/PoleEmploiSending.js';
+import * as monitoringTools from '../../../src/shared/infrastructure/monitoring-tools.js';
 import { PoleEmploiPayload } from '../../infrastructure/externals/pole-emploi/PoleEmploiPayload.js';
 import * as httpErrorsHelper from '../../infrastructure/http/errors-helper.js';
 import { httpAgent } from '../../infrastructure/http/http-agent.js';
-import * as monitoringTools from '../../infrastructure/monitoring-tools.js';
 
 const sendSharedParticipationResultsToPoleEmploi = async ({
   campaignParticipationId,
