+++ api/src/prescription/learner-management/infrastructure/utils/xml/detect-encoding.js
@@ -3,8 +3,8 @@ import * as fs from 'node:fs/promises';
 
 import xmlBufferTostring from 'xml-buffer-tostring';
 
-import { logErrorWithCorrelationIds } from '../../../../../../lib/infrastructure/monitoring-tools.js';
 import { FileValidationError } from '../../../../../shared/domain/errors.js';
+import { logErrorWithCorrelationIds } from '../../../../../shared/infrastructure/monitoring-tools.js';
 
 const { xmlEncoding } = xmlBufferTostring;
 const { Buffer } = buffer;
