+++ api/src/shared/infrastructure/plugins/pino.js
@@ -2,7 +2,7 @@ import crypto from 'node:crypto';
 
 import { stdSerializers } from 'pino';
 
-import { monitoringTools } from '../../../../lib/infrastructure/monitoring-tools.js';
+import { monitoringTools } from '../../../../src/shared/infrastructure/monitoring-tools.js';
 import { config } from '../../config.js';
 import { logger } from '../utils/logger.js';
