+++ api/tests/integration/infrastructure/plugins/pino_test.js
@@ -2,8 +2,8 @@ import { Writable } from 'node:stream';
 
 import pino from 'pino';
 
-import { monitoringTools } from '../../../../lib/infrastructure/monitoring-tools.js';
 import { config } from '../../../../src/shared/config.js';
+import { monitoringTools } from '../../../../src/shared/infrastructure/monitoring-tools.js';
 import * as pinoPlugin from '../../../../src/shared/infrastructure/plugins/pino.js';
 import { expect, generateValidRequestAuthorizationHeader, HttpTestServer, sinon } from '../../../test-helper.js';
