+++ api/tests/unit/infrastructure/http/http-agent_test.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 import { httpAgent } from '../../../../lib/infrastructure/http/http-agent.js';
-import { monitoringTools } from '../../../../lib/infrastructure/monitoring-tools.js';
+import { monitoringTools } from '../../../../src/shared/infrastructure/monitoring-tools.js';
 import { expect, sinon } from '../../../test-helper.js';
 
 const { post, get } = httpAgent;
